+++ backend/src/main/java/com/naren/movieticketbookingapplication/Dao/MovieDaoImpl.java
@@ -30,7 +30,7 @@ public class MovieDaoImpl implements MovieDao {
     }
 
     @Override
-    public Optional<Movie> getMovieById(Integer id) {
+    public Optional<Movie> getMovieById(Long id) {
         return movieRepository.findById(id);
     }
